login and signup refactor
