+++ src/components/Input/InputField.jsx
@@ -1,0 +1,17 @@
+function InputField({ type, name, value, onChange, placeholder, required = true }) {
+    return (
+      <input
+        type={type}
+        name={name}
+        value={value}
+        onChange={onChange}
+        placeholder={placeholder}
+        required={required}
+        className="w-full px-4 py-3 border border-gray-300 rounded-lg 
+                   focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+      />
+    );
+  }
+  
+  export default InputField;
+
