+++ src/components/Layouts/AuthLayout.jsx
@@ -1,0 +1,15 @@
+function AuthLayout({ title, subtitle, children }) {
+    return (
+      <div className="flex flex-col items-center justify-center h-screen bg-gradient-to-br from-blue-50 to-indigo-100">
+        <div className="bg-white rounded-2xl shadow-xl p-8 w-full max-w-md">
+          <div className="text-center mb-8">
+            <h2 className="text-3xl font-bold text-gray-900 mb-2">{title}</h2>
+            {subtitle && <p className="text-gray-600">{subtitle}</p>}
+          </div>
+          {children}
+        </div>
+      </div>
+    );
+  }
+  export default AuthLayout;
+
