--- src/components/Login.jsx
@@ -1,115 +1,0 @@
-import { useState } from "react";
-import { useNavigate } from "react-router-dom";
-import chatServices from "../main.service";
-
-function Login() {
-  const [formData, setFormData] = useState({
-    email: "",
-    password: ""
-  });
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState("");
-  const navigate = useNavigate();
-
-  const handleInputChange = (e) => {
-    const { name, value } = e.target;
-    setFormData(prev => ({
-      ...prev,
-      [name]: value
-    }));
-  };
-
-  const handleLogin = async (e) => {
-    e.preventDefault();
-    setLoading(true);
-    setError("");
-
-    try {
-      const response = await chatServices.loginService(formData);
-      console.log("login response---------", response);
-      if (response.status_code === 200 || response.message ) {
-        // Store user data in localStorage
-        localStorage.setItem('user', JSON.stringify(response.user));
-        localStorage.setItem('isAuthenticated', 'true');
-        
-        console.log('Login - User data stored:', response.user);
-        console.log('Login - Navigating to /chat');
-        
-        // Navigate to chat dashboard
-        navigate('/chat');
-      } else {
-        setError(response.message || "Login failed");
-      }
-    } catch (err) {
-      setError("An error occurred during login. Please try again.");
-      console.error("Login error:", err);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  return (
-    <div className="flex flex-col items-center justify-center h-screen bg-gradient-to-br from-blue-50 to-indigo-100">
-      <div className="bg-white rounded-2xl shadow-xl p-8 w-full max-w-md">
-        <div className="text-center mb-8">
-          <h2 className="text-3xl font-bold text-gray-900 mb-2">Welcome Back</h2>
-          <p className="text-gray-600">Sign in to your account</p>
-        </div>
-
-        <form onSubmit={handleLogin} className="space-y-6">
-          <div>
-            <input
-              type="email"
-              name="email"
-              required
-              className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-              placeholder="Email"
-              value={formData.email}
-              onChange={handleInputChange}
-            />
-          </div>
-          
-          <div>
-            <input
-              type="password"
-              name="password"
-              required
-              className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-              placeholder="Password"
-              value={formData.password}
-              onChange={handleInputChange}
-            />
-          </div>
-
-          {error && (
-            <div className="text-red-600 text-sm text-center bg-red-50 p-3 rounded-lg">
-              {error}
-            </div>
-          )}
-
-          <button
-            type="submit"
-            disabled={loading}
-            className="w-full bg-gradient-to-r from-blue-500 to-indigo-600 text-white py-3 rounded-lg hover:from-blue-600 hover:to-indigo-700 transition-all font-medium disabled:opacity-50 disabled:cursor-not-allowed"
-          >
-            {loading ? "Signing in..." : "Sign In"}
-          </button>
-        </form>
-
-        <div className="text-center mt-6">
-          <p className="text-gray-600">
-            Don't have an account?{" "}
-            <button 
-              onClick={() => navigate('/signup')} 
-              className="text-blue-600 hover:text-blue-700 font-medium"
-            >
-              Sign up
-            </button>
-          </p>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-export default Login;
