--- src/components/LoginPage.jsx
@@ -1,106 +1,0 @@
-import React, { useState } from 'react';
-import { MessageCircle} from 'lucide-react';
-import chatServices from '../main.service';
-
-const LoginPage = ({ onLogin, onSwitchToSignup }) => {
-    const [formData, setFormData] = useState({ email: '', password: '' });
-    const [loading, setLoading] = useState(false);
-    const [error, setError] = useState("");
-  
-    const handleSubmit = async (e) => {
-      e.preventDefault();
-      setLoading(true);
-      setError("");
-
-      try {
-        const response = await chatServices.loginService(formData);
-        
-        if (response.status_code === 200) {
-          // Store user data in localStorage
-          localStorage.setItem('user', JSON.stringify(response.user));
-          localStorage.setItem('isAuthenticated', 'true');
-          
-          // Call the onLogin callback with user data
-          onLogin({
-            id: response.user.id,
-            name: response.user.username || response.user.email,
-            email: response.user.email,
-            avatar: 'https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?w=150&h=150&fit=crop&crop=face'
-          });
-        } else {
-          setError(response.message || "Login failed");
-        }
-      } catch (err) {
-        setError("An error occurred during login. Please try again.");
-        console.error("Login error:", err);
-      } finally {
-        setLoading(false);
-      }
-    };
-  
-    return (
-      <div className="min-h-screen flex items-center justify-center bg-gradient-to-br from-blue-50 to-indigo-100 p-4">
-        <div className="bg-white rounded-2xl shadow-xl p-8 w-full max-w-md">
-          <div className="text-center mb-8">
-            <div className="bg-gradient-to-r from-blue-500 to-purple-600 w-16 h-16 rounded-full flex items-center justify-center mx-auto mb-4">
-              <MessageCircle className="h-8 w-8 text-white" />
-            </div>
-            <h1 className="text-3xl font-bold text-gray-900 mb-2">Welcome Back</h1>
-            <p className="text-gray-600">Sign in to continue to ChatApp</p>
-          </div>
-          
-          <form onSubmit={handleSubmit} className="space-y-6">
-            <div>
-              <label className="block text-sm font-medium text-gray-700 mb-2">Email</label>
-              <input
-                type="email"
-                required
-                className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500"
-                placeholder="Enter your email"
-                value={formData.email}
-                onChange={(e) => setFormData({...formData, email: e.target.value})}
-              />
-            </div>
-            
-            <div>
-              <label className="block text-sm font-medium text-gray-700 mb-2">Password</label>
-              <input
-                type="password"
-                required
-                className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500"
-                placeholder="Enter your password"
-                value={formData.password}
-                onChange={(e) => setFormData({...formData, password: e.target.value})}
-              />
-            </div>
-            
-            {error && (
-              <div className="text-red-600 text-sm text-center bg-red-50 p-3 rounded-lg">
-                {error}
-              </div>
-            )}
-
-            <button
-              type="submit"
-              disabled={loading}
-              className="w-full bg-gradient-to-r from-blue-500 to-purple-600 text-white py-3 rounded-lg hover:from-blue-600 hover:to-purple-700 transition-all font-medium disabled:opacity-50 disabled:cursor-not-allowed"
-            >
-              {loading ? "Signing in..." : "Sign In"}
-            </button>
-          </form>
-          
-          <div className="text-center mt-6">
-            <p className="text-gray-600">
-              Don't have an account?{' '}
-              <button onClick={onSwitchToSignup} className="text-blue-600 hover:text-blue-700 font-medium">
-                Sign up
-              </button>
-            </p>
-          </div>
-        </div>
-      </div>
-    );
-  };
-
-
-  export default LoginPage;
