--- src/components/SignupPage.jsx
@@ -1,131 +1,0 @@
-import React, { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { User } from 'lucide-react';
-import chatServices from '../main.service';
-
-const SignupPage = () => {
-    const navigate = useNavigate();
-    const [formData, setFormData] = useState({ 
-      username: '', 
-      email: '', 
-      password: '' 
-    });
-    const [loading, setLoading] = useState(false);
-    const [error, setError] = useState("");
-    const [success, setSuccess] = useState("");
-  
-    const handleInputChange = (e) => {
-      const { name, value } = e.target;
-      setFormData(prev => ({
-        ...prev,
-        [name]: value
-      }));
-    };
-
-    const handleSubmit = async (e) => {
-      e.preventDefault();
-      setLoading(true);
-      setError("");
-      setSuccess("");
-
-      try {
-        const response = await chatServices.signupService(formData);
-        console.log("signup response---------", response);
-        if (response.status_code === 200) {
-          setSuccess("Account created successfully! Please sign in.");
-          // Clear form
-          setFormData({ username: '', email: '', password: '' });
-          // Redirect to login after 2 seconds
-          setTimeout(() => {
-            navigate('/login');
-          }, 2000);
-        } else {
-          setError(response.message || "Signup failed");
-        }
-      } catch (err) {
-        setError("An error occurred during signup. Please try again.");
-        console.error("Signup error:", err);
-      } finally {
-        setLoading(false);
-      }
-    };
-  
-    return (
-      <div className="min-h-screen flex items-center justify-center bg-gradient-to-br from-green-50 to-blue-100 p-4">
-        <div className="bg-white rounded-2xl shadow-xl p-8 w-full max-w-md">
-          <div className="text-center mb-8">
-            <div className="bg-gradient-to-r from-green-500 to-blue-600 w-16 h-16 rounded-full flex items-center justify-center mx-auto mb-4">
-              <User className="h-8 w-8 text-white" />
-            </div>
-            <h1 className="text-3xl font-bold text-gray-900 mb-2">Create Account</h1>
-            <p className="text-gray-600">Join ChatApp today</p>
-          </div>
-          
-          <form onSubmit={handleSubmit} className="space-y-6">
-            <input
-              type="text"
-              name="username"
-              required
-              className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-green-500 focus:border-transparent"
-              placeholder="Username"
-              value={formData.username}
-              onChange={handleInputChange}
-            />
-            <input
-              type="email"
-              name="email"
-              required
-              className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-green-500 focus:border-transparent"
-              placeholder="Email"
-              value={formData.email}
-              onChange={handleInputChange}
-            />
-            <input
-              type="password"
-              name="password"
-              required
-              className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-green-500 focus:border-transparent"
-              placeholder="Password"
-              value={formData.password}
-              onChange={handleInputChange}
-            />
-
-            {error && (
-              <div className="text-red-600 text-sm text-center bg-red-50 p-3 rounded-lg">
-                {error}
-              </div>
-            )}
-
-            {success && (
-              <div className="text-green-600 text-sm text-center bg-green-50 p-3 rounded-lg">
-                {success}
-              </div>
-            )}
-            
-            <button
-              type="submit"
-              disabled={loading}
-              className="w-full bg-gradient-to-r from-green-500 to-blue-600 text-white py-3 rounded-lg hover:from-green-600 hover:to-blue-700 transition-all font-medium disabled:opacity-50 disabled:cursor-not-allowed"
-            >
-              {loading ? "Creating Account..." : "Create Account"}
-            </button>
-          </form>
-          
-          <div className="text-center mt-6">
-            <p className="text-gray-600">
-              Already have an account?{' '}
-              <button 
-                onClick={() => navigate('/login')} 
-                className="text-green-600 hover:text-green-700 font-medium"
-              >
-                Sign in
-              </button>
-            </p>
-          </div>
-        </div>
-      </div>
-    );
-  };
-
-
-  export default SignupPage;
